+++ docs/releasenotes.md
@@ -13,11 +13,12 @@ Packages are available on NuGet: [protobuf-net](https://www.nuget.org/packages/p
 
 ## pending
 
+3.0.131
+
 - support unknwon/extension fields on models that involve inheritance (via either `Extensible` or `ITypedExtensible`)
 - detect Google.Protobuf types and provide guidance (#722)
 - don't throw if `EnumPassthru` is explicitly set to `true` (#881) from code compiled against v2
-- fix #479
-
+- fix #479 (also backported as 2.4.7)
 
 ## 3.0.62